Move gomega matchers to pkg/utils/test/matchers
