+++ pkg/utils/test/matchers/deep_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gomega_test
+package matchers_test
 
 import (
 	corev1 "k8s.io/api/core/v1"
@@ -21,7 +21,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	. "github.com/gardener/gardener/test/gomega"
+	. "github.com/gardener/gardener/pkg/utils/test/matchers"
 )
 
 var _ = Describe("Deep", func() {
