+++ pkg/utils/test/matchers/matchers.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gomega
+package matchers
 
 import (
 	"github.com/onsi/gomega/format"
